Create Storage.java to load and save tasks
